refactor: simplify the card dialogs by removing open state2
