+++ src/app/(auth)/okr/[id]/components/OkrCard/AddKeyResultDialog.tsx
@@ -4,25 +4,26 @@ import {
   DialogFooter,
   DialogHeader,
   DialogTitle,
+  DialogTrigger,
 } from "@/components/ui/dialog";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
+import { DialogClose } from "@radix-ui/react-dialog";
 
 export default function AddKeyResultDialog({
-  open,
-  closeDialog,
+  children,
   addKeyResult,
 }: {
-  open: boolean;
-  closeDialog: () => void;
+  children: React.ReactNode;
   addKeyResult: (keyResult: string) => void;
 }) {
   const [keyResult, setKeyResult] = useState<string>("");
 
   return (
-    <Dialog open={open} onOpenChange={closeDialog}>
+    <Dialog>
+      <DialogTrigger asChild>{children}</DialogTrigger>
       <DialogContent>
         <DialogHeader>
           <DialogTitle>Add Key Result</DialogTitle>
@@ -38,14 +39,11 @@ export default function AddKeyResultDialog({
           </div>
         </div>
         <DialogFooter>
-          <Button
-            onClick={() => {
-              addKeyResult(keyResult);
-              closeDialog();
-            }}
-          >
-            Add Key Result
-          </Button>
+          <DialogClose>
+            <Button onClick={() => addKeyResult(keyResult)}>
+              Add Key Result
+            </Button>
+          </DialogClose>
         </DialogFooter>
       </DialogContent>
     </Dialog>
